Input error accessment
Accessing user input type error by creating two functions which are the getInt and getFloat function
these functions force the user to enter the appropriate datatype when invoked.

+++ src/com/company/Student.java
@@ -3,10 +3,10 @@ package com.company;
 public class Student {
     String name;
     Integer age;
-    Integer englishScore;
-    Integer mathScore;
-    Integer average;
-    public Student(String _name, Integer _age, Integer _englishScore, Integer _mathScore){
+    Float englishScore;
+    Float mathScore;
+    Float average;
+    public Student(String _name, Integer _age, Float _englishScore, Float _mathScore){
         name = _name;
         age = _age;
         englishScore = _englishScore;
